// NextSpring/frontend/src/lib/backend/apiV1/schema.d.ts
export interface components {
    schemas: {
// ...
        PageDtoPostDto: {
            pageable: components["schemas"]["PageableDto"];
            content: components["schemas"]["PostDto"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}